backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
Flask==2.2.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.3.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.0.2
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
setuptools==75.1.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==2.2.0
bcrypt==3.2.0
pyjwt==2.1.0
